+++ components/layout/fragment_tree/fragment.rs
@@ -23,7 +23,6 @@ use super::{
 use crate::SharedStyle;
 use crate::cell::ArcRefCell;
 use crate::flow::inline::line::TextRunOffsets;
-use crate::fragment_tree::ContainingBlockCalculation;
 use crate::geom::{LogicalSides, PhysicalPoint, PhysicalRect};
 use crate::layout_impl::LayoutThread;
 use crate::style_ext::ComputedValuesExt;
@@ -500,5 +499,39 @@ impl CollapsedMargin {
 
     pub fn solve(&self) -> Au {
         self.max_positive + self.min_negative
+    }
+}
+
+/// A token which ensures the calculation and assignment of cumulative containing
+/// blocks to fragments in the fragment tree. This is used because these cumulative
+/// containing block offsets are set during stacking context tree construction, but
+/// some queries might need them beforehand. If the query is executed before stacking
+/// context tree construction, a quick traversal is performed to calculate them for
+/// the purpose of the query.
+pub(crate) enum ContainingBlockCalculation<'a> {
+    /// This token variant is for the purpose of a layout query. In this case, if stacking
+    /// context tree construction has not yet taken place, a cumulative containing block
+    /// calculation traversal will be performed.
+    Lazy { layout_thread: &'a LayoutThread },
+    /// This token variant is used when the code can guarantee that stacking context
+    /// tree construction has already taken place.
+    ///
+    /// Note: Using this before stacking context tree construction can lead
+    /// to incorrect layout or layout query results!
+    AlreadyDoneWithStackingContextTree,
+}
+
+impl ContainingBlockCalculation<'_> {
+    pub(crate) fn ensure(&self) {
+        match self {
+            Self::Lazy { layout_thread } => layout_thread.ensure_containing_block_calculation(),
+            Self::AlreadyDoneWithStackingContextTree => {},
+        }
+    }
+}
+
+impl<'a> From<&'a LayoutThread> for ContainingBlockCalculation<'a> {
+    fn from(layout_thread: &'a LayoutThread) -> Self {
+        Self::Lazy { layout_thread }
     }
 }
